Fixed Gradle Clojure build by downgrading Gradle to 3.5.1.
It seems the Nebula Clojure plugin is not really compatible with Gradle 4+ (?) as it
crashes calling a non-existing constructor. Furthermore I did a little bit of
cleanup in the build script. The application can be run from Gradle through

./gradlew run

This fails currently in the main() method, but this is beyond my knowledge of how Java-Clojure
interop works. It seems it cannot find the method called on the generated class.